requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-json-logger>=2.0.0
streamlit-option-menu>=0.3.0
streamlit-sortables>=0.3.0
openpyxl>=3.1.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
httplib2>=0.22.0
